requirements.txt
langchain
langchain-community
faiss-cpu
tiktoken
sentence-transformers
streamlit
python-dotenv
pandas
trafilatura
torch
transformers
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.40.2
trl==0.4.7
streamlit_server_state
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery[pandas] 
langchain-google-vertexai